[#151-REFACTOR] fixed caption
