+++ src/app/components/home/game-slider-item/game-slider-item.component.html
@@ -21,7 +21,7 @@
       </div>
     </div>
     <div class="flex flex-col gap-1">
-      <h1 class="text-[9px] font-thin">{{ captionForGameType | uppercase }}</h1>
+      <h1 class="text-[9px] font-thin">{{ caption().gameType[data().type] | uppercase }}</h1>
       <h2 class="text-sm">{{ data().name }}</h2>
       <div>
         @if (data().isFree) {
