+++ src/app/components/home/game-slider-item/game-slider-item.component.ts
@@ -3,7 +3,6 @@ import {GameSliderItemInputModel} from "../models/game-slider-item-input.model";
 import {SizeEnum} from "../enums/size.enum";
 import {PriceLabelModel} from "../models/price-label.model";
 import {GameSliderCaptionModel} from "../models/caption-models/game-slider-caption.model";
-import {GameType} from "../enums/game-type.enum";
 
 @Component({
     selector: 'app-game-slider-item',
@@ -21,32 +20,17 @@ export class GameSliderItemComponent implements OnInit {
 
     public readonly SizeEnum = SizeEnum;
     public priceLabelData: PriceLabelModel | undefined;
-    public captionForGameType: string | undefined;
 
     //endregion
 
     //region Lifecycle methods
     ngOnInit(): void {
-        this.captionForGameType = this._getCaptionForType(this.data().type);
-
         this._setPriceLabelData();
     }
 
     //endregion
 
     //region Main logic methods
-    private _getCaptionForType(type: GameType): string {
-        switch (type) {
-            case GameType.BASE_GAME:
-                return this.caption().gameType.BASE_GAME;
-            case GameType.EDITION:
-                return this.caption().gameType.EDITION;
-
-            default:
-                return this.caption().gameType.BASE_GAME;
-        }
-    }
-
     private _setPriceLabelData(): void {
         this.priceLabelData = this._convertGameSliderItemInputModelToPriceLabelModel(this.data());
     }
