+++ src/app/components/home/models/caption-models/game-slider-caption.model.ts
@@ -1,4 +1,4 @@
-import {GameTypeCaptionModel} from "../enum-caption-model/game-type-caption.model";
+import {GameTypeCaptionModel} from "../../../../models/enum-caption-model/game-type-caption.model";
 
 export interface GameSliderCaptionModel {
   freeTitle: string;
