--- src/app/components/home/models/enum-caption-model/game-type-caption.model.ts
@@ -1,4 +1,0 @@
-export interface GameTypeCaptionModel {
-  BASE_GAME: string;
-  EDITION: string;
-}
