+++ src/app/enums/game-type.enum.ts
