+++ src/app/models/enum-caption-model/game-type-caption.model.ts
@@ -1,0 +1,4 @@
+export interface GameTypeCaptionModel {
+  'BASE GAME': string;
+  'EDITION': string;
+}
