+++ src/assets/i18n/en.json
@@ -29,12 +29,12 @@
     },
     "GameSliderItem": {
       "freeTitle": "Free"
-    },
-    "enum-captions": {
-      "gameType": {
-        "BASE_GAME": "Base Game",
-        "EDITION": "Edition"
-      }
+    }
+  },
+  "enum-captions": {
+    "gameType": {
+      "BASE_GAME": "Base Game",
+      "EDITION": "Edition"
     }
   }
 }
